bug fixes and code seperation. Continuing work on multi sampling
