+++ src/blop/gradient/__init__.py
@@ -1,3 +1,4 @@
-from .Scipy import SCP, Scipy, ScipyCFG, ScipyOptimizer
+from .optimizer import SCP, ScipyCFG, ScipyOptimizer
+from .Scipy import Scipy
 
 __all__ = ["SCP", "ScipyCFG", "Scipy", "ScipyOptimizer"]
